Lagt til iverksetting av barnetilsyn (#1379)

+++ src/main/kotlin/no/nav/familie/ef/sak/vedtak/domain/Vedtak.kt
@@ -52,11 +52,11 @@ data class Barnetilsynperiode(val datoFra: LocalDate,
 
 data class PeriodeMedBeløp(val datoFra: LocalDate,
                            val datoTil: LocalDate,
-                           val beløp: BigDecimal) {
+                           val beløp: Int) {
 
     fun tilDto() = PeriodeMedBeløpDto(årMånedFra = YearMonth.from(datoFra),
-                                           årMånedTil = YearMonth.from(datoTil),
-                                           beløp = beløp)
+                                      årMånedTil = YearMonth.from(datoTil),
+                                      beløp = beløp)
 }
 
 data class PeriodeWrapper(val perioder: List<Vedtaksperiode>)
